requirements.txt
Django==1.11.3
pytz==2017.2
djangorestframework=3.6.3
